updated EffectiveArea in ssSelections.  added fastJetEffArea0N_vM (N=03 or 04, M=1 or 2) which the old (v1) and new (v2) effective ares, modified some electron isolations functions to use this so that the conesize option actually has some meaning. fixed many warnings in various pieces

+++ jetSelections.cc
@@ -1,4 +1,4 @@
-// $Id: jetSelections.cc,v 1.39 2013/01/19 20:48:44 benhoob Exp $
+// $Id: jetSelections.cc,v 1.40 2013/01/28 23:44:26 kelley Exp $
 
 #include <algorithm>
 #include <utility>
@@ -487,7 +487,7 @@ float pfjet_beta(int ijet, int power , float dzcut , int ivtx , bool verbose ) {
 
     int ican = cands[ivc];
 
-    if( ican > cms2.pfcands_charge().size() ){
+    if( ican > static_cast<int>(cms2.pfcands_charge().size()) ){
       cout << __FILE__ << " " << __LINE__ << " ERROR! ican, numCandidates " << ican << ", " << cms2.pfcands_charge().size() << endl;
       return -3.0;
     }
